Divide the main.dart into 5 subpages and update the dependencies
