+++ web_app/lib/data_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class DataPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Data Page'),
+    );
+  }
+}
