+++ web_app/lib/diary_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class DiaryPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Diary Page'),
+    );
+  }
+}
