+++ web_app/lib/info_page.dart
