+++ web_app/lib/setting_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class SettingPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Setting Page'),
+    );
+  }
+}
