+++ web_app/pubspec.yaml
@@ -1,11 +1,11 @@
 name: web_app
 description: "A new Flutter project."
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 
 version: 0.0.1+1
 
 environment:
-  sdk: ^3.1.1
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -13,17 +13,21 @@ dependencies:
   table_calendar: ^3.1.1
   shared_preferences: ^2.1.0
   english_words: ^4.0.0
-<<<<<<< HEAD
   provider: ^6.1.0
-=======
-  provider: ^6.0.0
->>>>>>> main
   bulleted_list: ^0.0.1+0.1a
+  intl: ^0.19.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.3
+  flutter_lints: ^4.0.0
+  leak_tracker: ^10.0.4
+  leak_tracker_flutter_testing: ^3.0.3
+  lints: ^4.0.0
+  material_color_utilities: ^0.8.0
+  meta: ^1.12.0
+  test_api: ^0.7.0
+  vm_service: ^14.2.1
 
 flutter:
   uses-material-design: true
